moved components to folders and added export index
