+++ clo-workflows/Scripts/src/component/Client/index.ts
@@ -1,0 +1,2 @@
+import Client from "./Client"
+export default Client
