+++ clo-workflows/Scripts/src/component/DescriptiveCheckbox/index.ts
@@ -1,0 +1,2 @@
+import DescriptiveCheckbox from "./DescriptiveCheckbox"
+export default DescriptiveCheckbox
