+++ clo-workflows/Scripts/src/component/DescriptiveDropdown/index.ts
@@ -1,0 +1,2 @@
+import DescriptiveDropdown from "./DescriptiveDropdown"
+export default DescriptiveDropdown
