+++ clo-workflows/Scripts/src/component/Employee/index.ts
@@ -1,0 +1,2 @@
+import Employee from "./Employee"
+export default Employee
