+++ clo-workflows/Scripts/src/component/EmployeeDashboard/index.ts
@@ -1,0 +1,2 @@
+import EmployeeDashboard from "./EmployeeDashboard"
+export default EmployeeDashboard
