+++ clo-workflows/Scripts/src/component/FormControlGroup/index.ts
@@ -1,0 +1,2 @@
+import FormControlGroup from "./FormControlGroup"
+export default FormControlGroup
