+++ clo-workflows/Scripts/src/component/HeaderBreadcrumb/index.ts
@@ -1,0 +1,2 @@
+import HeaderBreadcrumb from "./HeaderBreadcrumb"
+export default HeaderBreadcrumb
