+++ clo-workflows/Scripts/src/component/LoadingPage/index.ts
@@ -1,0 +1,2 @@
+import LoadingPage from "./LoadingPage"
+export default LoadingPage
