+++ clo-workflows/Scripts/src/component/Message/index.ts
@@ -1,0 +1,2 @@
+import Message from "./Message"
+export default Message
