+++ clo-workflows/Scripts/src/component/NonScrollableList/index.ts
@@ -1,0 +1,2 @@
+import NonScrollableList from "./NonScrollableList"
+export default NonScrollableList
