+++ clo-workflows/Scripts/src/component/NotesBox/index.ts
@@ -1,0 +1,2 @@
+import NotesBox from "./NotesBox"
+export default NotesBox
