+++ clo-workflows/Scripts/src/component/ProcessBriefList/index.ts
@@ -1,0 +1,2 @@
+import ProcessBriefList from "./ProcessBriefList"
+export default ProcessBriefList
