+++ clo-workflows/Scripts/src/component/ProcessDetail/index.ts
@@ -1,0 +1,2 @@
+import ProcessDetail from "./ProcessDetail"
+export default ProcessDetail
