+++ clo-workflows/Scripts/src/component/ProcessFormModal/index.ts
@@ -1,0 +1,2 @@
+import ProcessFormModal from "./ProcessFormModal"
+export default ProcessFormModal
