+++ clo-workflows/Scripts/src/component/ProjectDetail/index.ts
@@ -1,0 +1,2 @@
+import ProjectDetail from "./ProjectDetail"
+export default ProjectDetail
