+++ clo-workflows/Scripts/src/component/ProjectFormModal/index.ts
@@ -1,0 +1,2 @@
+import ProjectFormModal from "./ProjectFormModal"
+export default ProjectFormModal
