+++ clo-workflows/Scripts/src/component/ProjectProcessList/index.ts
@@ -1,0 +1,2 @@
+import ProjectProcessList from "./ProjectProcessList"
+export default ProjectProcessList
