+++ clo-workflows/Scripts/src/component/RoleSteps/index.ts
@@ -1,0 +1,2 @@
+import RoleSteps from "./RoleSteps"
+export default RoleSteps
