+++ clo-workflows/Scripts/src/component/WorkDetail/WorkDetail.tsx
@@ -5,7 +5,7 @@ import FormControlGroup from "./FormControlGroup"
 import { NotesBox } from "./NotesBox"
 import { PrimaryButton, IconButton } from "office-ui-fabric-react/lib/Button"
 import { SessionStore } from "../store/SessionStore"
-import { NoteSource } from "../model/Note"
+import { NoteScope, NoteSource } from "../model/Note"
 
 const wrapperStyle = {
     padding: "20 0",
@@ -28,17 +28,18 @@ const titleStlyes = {
 const submitButtonStlyes = {
     display: "flex",
     justifyContent: "center",
-    marginTop: 30
+    marginTop: 30,
+    marginRight: 25
 } as React.CSSProperties
 
-const projectHeaderStyles = { display: "flex" }
 const editButtonStyles = { transform: "translateX(-25px)" }
+const workHeaderStyles = { display: "flex" } as React.CSSProperties
 const formColumnStyles = { padding: "0 8 0 30" }
 const notesColumnStyles = { padding: "0 30 0 8" }
 
 @inject("rootStore")
 @observer
-export class ProjectDetail extends React.Component<any, any> {
+export default class WorkDetail extends React.Component<any, any> {
 
     public componentWillMount() {
         this.employeeStore = this.props.rootStore.employeeStore
@@ -52,46 +53,47 @@ export class ProjectDetail extends React.Component<any, any> {
         return (
             <div style={wrapperStyle}>
                 <div style={formColumnStyles}>
-                    <div style={projectHeaderStyles}>
-                        <div style={titleStlyes}>{this.employeeStore.canEditSelectedProject ? "Edit Project" : "View Project"}</div>
+                    <div style={workHeaderStyles}>
+                        <div style={titleStlyes}>{this.employeeStore.canEditSelectedWork ? "Edit Work" : "View Work"}</div>
                         <div style={editButtonStyles}>
                             <IconButton
                                 disabled={!this.employeeStore.isSelectedRequestActive}
-                                iconProps={ this.employeeStore.canEditSelectedProject ? {iconName: "BoxMultiplySolid"} : {iconName: "edit"} }
-                                onClick={this.employeeStore.canEditSelectedProject
-                                    ? this.employeeStore.stopEditingSelectedProject
-                                    : this.employeeStore.startEditingSelectedProject
+                                iconProps={ this.employeeStore.canEditSelectedWork ? {iconName: "BoxMultiplySolid"} : {iconName: "edit"} }
+                                onClick={this.employeeStore.canEditSelectedWork
+                                    ? this.employeeStore.stopEditingSelectedWork
+                                    : this.employeeStore.startEditingSelectedWork
                                 }
                             />
                         </div>
                     </div>
                     <FormControlGroup
-                        data={this.employeeStore.selectedProject}
-                        formControls={this.employeeStore.selectedProjectView.formControls}
-                        updateFormField={this.employeeStore.updateSelectedProject}
-                        validation={this.employeeStore.selectedProjectValidation}
+                        data={this.employeeStore.selectedWork}
+                        formControls={this.employeeStore.selectedWorkView.formControls}
+                        updateFormField={this.employeeStore.updateSelectedWork}
+                        validation={this.employeeStore.selectedWorkValidation}
                         width={350}
                     />
                     {
-                        this.employeeStore.canEditSelectedProject &&
+                        this.employeeStore.canEditSelectedWork &&
                         <div style={submitButtonStlyes}>
                             <PrimaryButton text="Submit Changes"
-                                onClick={this.employeeStore.submitSelectedProject}
-                                disabled={!this.employeeStore.canSubmitSelectedProject}
+                                onClick={this.employeeStore.submitSelectedWork}
+                                disabled={!this.employeeStore.canSubmitSelectedWork}
                             />
                         </div>
                     }
                 </div>
                 <div style={notesColumnStyles}>
                     <NotesBox
-                        title="Project Notes"
-                        notes={this.employeeStore.selectedProjectNotes}
+                        title="Work Notes"
+                        notes={this.employeeStore.selectedWorkNotes}
                         onCreateNote={this.employeeStore.submitNewNote}
-                        onDeleteNote={this.employeeStore.deleteNote}
                         onUpdateNote={this.employeeStore.updateNote}
-                        disableButtons={this.employeeStore.asyncPendingLockout || !this.employeeStore.isSelectedRequestActive}
+                        onDeleteNote={this.employeeStore.deleteNote}
                         currentUser={this.sessionStore.currentUser}
-                        noteSource={NoteSource.PROJECT}
+                        disableButtons={this.employeeStore.asyncPendingLockout || !this.employeeStore.isSelectedRequestActive}
+                        maxScope={NoteScope.EMPLOYEE}
+                        noteSource={NoteSource.WORK}
                     />
                 </div>
             </div>
