+++ clo-workflows/Scripts/src/component/WorkDetail/index.ts
@@ -1,0 +1,2 @@
+import WorkDetail from "./WorkDetail"
+export default WorkDetail
