+++ clo-workflows/Scripts/src/component/WorkForm/index.ts
@@ -1,0 +1,2 @@
+import WorkForm from "./WorkForm"
+export default WorkForm
